pyproject.toml
[tool.poetry]
name = "reviews-parsing-mlsys"
version = "0.1.0"
description = ""
authors = ["Quy Dinh <quy.dinh3195@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# Required by mlserver
python = ">=3.9,<3.12"
pandas = "^2.2.2"
scikit-learn = "*"
openai = "^1.35.7"
python-dotenv = "^1.0.1"

# Download HuggingFace datasets via Pandas API
pyarrow = "*"
# Issue with load_dataset so need to fix fsspec: https://stackoverflow.com/a/77433141/7456751
fsspec = "2023.9.2"
huggingface-hub = "^0.23.4"
# HuggingFace datasets
datasets = "^2.16"

# spacy is used to convert to IOB2 format
spacy = "^3.7.5"

# Needed for tutorial notebook: https://huggingface.co/learn/nlp-course/en/chapter7/2
evaluate = "^0.4.2"
# At July 10 we are suggested to use version 4.41.0 at max to avoid compatibility issues with MLflow
# Ref: The 'transformers' MLflow Models integration is known to be compatible with the following package version ranges: ``4.25.1`` -  ``4.41.0``. MLflow Models integrations with transformers may not succeed when used with package versions outside of this range.
transformers = {extras = ["sentencepiece"], version = "4.41.0"}
accelerate = "^0.32.0"
seqeval = "^1.2.2"
ollama = "^0.2.1"
loguru = "^0.7.2"

# MLflow
mlflow = "2.14.2"
# Somehow needed for logging Trasnformers model to MLflow 
torchvision = "*"
mlserver = "^1.6.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-jaeger = "^1.21.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.3"
ipywidgets = "^8.1.3"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
